+++ src/components/PeriodSelect/index.less
@@ -11,4 +11,8 @@
     margin-top: 10px;
     width: 100%;
   }
+  .period-title {
+    font-size: 14px;
+    margin-bottom: 15px;
+  }
 }
